Separated triggering of db update function to separate file.
