+++ package.json
@@ -13,9 +13,9 @@
     "update_database": "node src/services/database_updater/database_updater.js",
     "update_studentnumbers_dev": "NODE_ENV=dev node src/services/database_updater/student_list_updater.js",
     "init_redis": "node src/util/init_redis.js",
+    "update_database_dev_new_api": "NODE_ENV=dev node src/services/doo_api_database_updater/run_updater.js",
     "seed_redis": "node src/util/add_json_duplicates_to_redis.js && node src/util/map_course_codes.js",
     "studentnumbers_to_file": "NODE_ENV=dev node src/services/doo_api_database_updater/db_studentnumbers_to_file.js",
-    "update_database_dev_new_api": "NODE_ENV=dev node src/services/doo_api_database_updater/database_updater.js",
     "lint": "eslint --parser babel-eslint src",
     "coverage": "NODE_ENV=test nyc ava"
   },
